Moved WordPress traits to `Native` NS
