+++ src/WordPress/Native/WpInsertCapableTrait.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace RebelCode\Storage\Resource\WordPress;
+namespace RebelCode\Storage\Resource\WordPress\Native;
 
 use Dhii\Util\String\StringableInterface as Stringable;
 use InvalidArgumentException;
